// src/clr-angular/utils/focus/key-codes.enum.ts
/*
 * Copyright (c) 2016-2019 VMware, Inc. All Rights Reserved.
 * This software is released under MIT license.
 * The full license information can be found in LICENSE in the root directory of this project.
 */

// Event.key key codes for standard browsers and IE/Edge

export enum KeyCodes {
  ArrowUp = 'ArrowUp',
  ArrowDown = 'ArrowDown',
  ArrowRight = 'ArrowRight',
  ArrowLeft = 'ArrowLeft',
  Space = ' ',
  Escape = 'Escape',
  Home = 'Home',
  End = 'End',
  Enter = 'Enter',
  Tab = 'Tab',
}

export enum IEKeyCodes {
  ArrowUp = 'Up',
  ArrowDown = 'Down',
  ArrowRight = 'Right',
  ArrowLeft = 'Left',
  Space = 'Spacebar',
  Escape = 'Esc',
  Home = 'Home',
  End = 'End',
  Enter = 'Enter',
  Tab = 'Tab',
}